+++ tinfoil.cabal
@@ -34,6 +34,8 @@ library
                        Paths_ambiata_tinfoil
                        Tinfoil
                        Tinfoil.Data
+                       Tinfoil.Data.KDF
+                       Tinfoil.Data.Random
                        Tinfoil.KDF
                        Tinfoil.KDF.Scrypt
                        Tinfoil.KDF.Scrypt.Internal
@@ -145,6 +147,3 @@ benchmark bench
                      , random                          == 1.1.*
                      , semigroups
                      , text                            == 1.2.*
-
-
-
